local-telegram-unzip-bot/requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
python-dotenv>=1.0,<2
aiohttp>=3.9,<4
